requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
black==26.3.1
cffi==2.0.0
click==8.3.1
cryptography==46.0.6
fastapi==0.135.2
greenlet==3.3.2
h11==0.16.0
idna==3.11
mypy_extensions==1.1.0
packaging==26.0
pathspec==1.0.4
platformdirs==4.9.4
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
PyJWT==2.12.1
PyMySQL==1.1.2
python-decouple==3.8
python-dotenv==1.2.2
pytokens==0.4.1
SQLAlchemy==2.0.48
starlette==1.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.42.0
websockets==12.0
